Move xmlHttpRequest to bridge.js
