+++ src/content-scripts/loader.js
@@ -1,6 +1,7 @@
 //@license magnet:?xt=urn:btih:1f739d935676111cfff4b4693e3816e664797050&dn=gpl-3.0.txt GPL-v3
 import { GM } from "./gm-api";
-import { inject, bridgeAction, IITCButtonInitJS } from "./utils";
+import { inject, IITCButtonInitJS } from "./utils";
+import { bridgeAction } from "@/content-scripts/bridge";
 
 function preparePage() {
   document.addEventListener("DOMContentLoaded", function() {
